pyproject.toml
[project]
name = "vpn-registration"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.11",
]
